requirements.txt
pandas
ollama
huggingface_hub
google-genai
